+++ packages/facebook-oauth/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "Facebook OAuth flow",
-  version: "1.9.0-beta230.4"
+  version: "1.9.0-beta230.5"
 });
 
 Package.onUse(api => {