+++ packages/meteor-developer-config-ui/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: 'Blaze configuration templates for the Meteor developer accounts OAuth.',
-  version: '1.0.2-beta230.4'
+  version: '1.0.2-beta230.5'
 });
 
 Package.onUse(api => {