+++ packages/minimongo/package.js
@@ -1,6 +1,6 @@
 Package.describe({
   summary: "Meteor's client-side datastore: a port of MongoDB to Javascript",
-  version: '1.7.0-beta230.4'
+  version: '1.7.0-beta230.5'
 });
 
 Package.onUse(api => {